requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
dj-rest-auth==4.0.1
djangorestframework-simplejwt==5.2.2
django-allauth==0.54.0